refactor[#234]: action/log 패키지 생성
