+++ src/main/java/org/highfive/backend/action/log/ActionLog.java
@@ -1,10 +1,11 @@
-package org.highfive.backend.action;
+package org.highfive.backend.action.log;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+import org.highfive.backend.action.Action;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
